Update project logistics and organization. No code changes.
