+++ README.md
@@ -1,0 +1,8 @@
+# JV Setup Keithley Controller
+A GUI app to controller JV and MPPT experiments.
+
+## Use
+**TODO:** How to use the app? Add documentation here.
+
+## Building
+**TODO:** How to build the app into an executable?
